requirements.txt
streamlit==1.27.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python-headless==4.8.0.76
pillow==10.0.0
